+++ webapp/package.json
@@ -13,6 +13,7 @@
     "axios": "^1.7.9",
     "react": "^18.2.0",
     "react-awesome-button": "^7.0.5",
+    "react-background-slider": "^4.0.0",
     "react-custom-scrollbars-2": "^4.5.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^7.2.0",
