+++ webapp/src/Home.js
@@ -1,92 +1,45 @@
+
+///////////////////////////////////////////////////////
 import React, { useState, useCallback } from "react";
 import { Link } from "react-router-dom";
 import Nav from "./components/Nav";
 import Particles from "react-tsparticles";
-import { loadSlim } from "tsparticles-slim"; // Cambiado de loadFull a loadSlim
+import { loadSlim } from "tsparticles-slim";
+import BackgroundSlider from "react-background-slider"; // Importamos el nuevo slider
+
 import "./Home.css";
 
 function Home() {
     const [isLoggedIn, setIsLoggedIn] = useState(!!localStorage.getItem('token'));
-    
-    // Inicializar tsparticles con loadSlim en lugar de loadFull
-    const particlesInit = useCallback(async (engine) => {
-        await loadSlim(engine);  // Usar loadSlim para una versión más ligera
-    }, []);
+
+    // Lista de imágenes de fondo
+    const imagesList = [
+        "https://images.dog.ceo/breeds/labrador/n02099712_3503.jpg",
+        "https://images.dog.ceo/breeds/labrador/n02099712_5844.jpg",
+        "https://images.dog.ceo/breeds/labrador/n02099712_5343.jpg",
+        "https://images.dog.ceo/breeds/labrador/n02099712_7481.jpg",
+        "https://images.dog.ceo/breeds/labrador/n02099712_7414.jpg"
+    ];
 
     return (
-        <div style={{ height: "100vh", display: "flex", flexDirection: "column" }}>
+        <div style={{ height: "100vh", display: "flex", flexDirection: "column", position: "relative" }}>
             <Nav />
-            
-            {/* Configuración de partículas */}
-            <Particles
-                id="tsparticles"
-                init={particlesInit}
-                style={{
-                    position: "absolute",
-                    width: "100%",
-                    height: "100%",
-                    zIndex: 0
-                }}
-                options={{
-                    background: {
-                        color: {
-                            value: "transparent"
-                        }
-                    },
-                    fpsLimit: 60,
-                    particles: {
-                        color: {
-                            value: "#8590AA"
-                        },
-                        links: {
-                            color: "#8590AA",
-                            distance: 150,
-                            enable: true,
-                            opacity: 0.7,     // Aumentado de 0.5 a 0.7
-                            width: 2.5        // Aumentado de 1 a 2.5
-                        },
-                        move: {
-                            enable: true,
-                            direction: "none",
-                            outModes: {
-                                default: "bounce"
-                            },
-                            random: false,
-                            speed: 1,
-                            straight: false
-                        },
-                        number: {
-                            density: {
-                                enable: true,
-                                area: 800
-                            },
-                            value: 80
-                        },
-                        opacity: {
-                            value: 0.8       // Aumentado de 0.5 a 0.8
-                        },
-                        shape: {
-                            type: "circle"
-                        },
-                        size: {
-                            value: { min: 3, max: 6 }  // Aumentado de {min: 1, max: 3} a {min: 3, max: 6}
-                        }
-                    },
-                    detectRetina: true
-                }}
+
+            {/* Fondo Animado */}
+            <BackgroundSlider
+                images={imagesList}
+                duration={2}  // Duración de cada imagen en segundos
+                transition={1} // Tiempo de transición entre imágenes en segundos
             />
-            
-            <div
-                className="container"
-                style={{
-                    flex: 1,
-                    backgroundSize: "cover",
-                    backgroundPosition: "center",
-                    backgroundRepeat: "no-repeat",
-                    position: "relative",
-                    zIndex: 1  // Asegurarse de que esté por encima de las partículas
-                }}
-            >
+
+            <div className="container" style={{
+                flex: 1,
+                position: "relative",
+                zIndex: 1,  // Asegura que el contenido esté por encima del fondo
+                display: "flex",
+                justifyContent: "center",
+                alignItems: "center"
+            }}>
                 {!isLoggedIn ? (
                     <button className="play-button">
                         <Link to="/login" style={{ color: "white" }}>
