+++ webapp/src/components/Login.js
@@ -1,9 +1,12 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 import { TextField, Button, Typography, Box, CircularProgress } from '@mui/material';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
 import { Link } from "react-router-dom";
 import "./LoginRegister.css";
+
+import Particles from "react-tsparticles";
+import { loadSlim } from "tsparticles-slim"; // Cambiado de loadFull a loadSlim
 
 function Login() {
     const navigate = useNavigate();
@@ -16,6 +19,12 @@ function Login() {
     const apiKey = process.env.REACT_APP_LLM_API_KEY || 'None';
    // const message = await axios.post(`${apiEndpoint}/askllm`, { question, model, apiKey })
    // setMessage(message.data.answer); left this as a reference for future use
+
+
+   // Inicializar tsparticles con loadSlim en lugar de loadFull
+       const particlesInit = useCallback(async (engine) => {
+           await loadSlim(engine);  // Usar loadSlim para una versión más ligera
+       }, []);
 
     useEffect(() => {
         if (isLoggedIn) {
@@ -50,6 +59,67 @@ function Login() {
     };
 
     return (
+        <>
+
+        {/* Configuración de partículas */}
+                    <Particles
+                        id="tsparticles"
+                        init={particlesInit}
+                        style={{
+                            position: "absolute",
+                            width: "100%",
+                            height: "100%",
+                            zIndex: 0
+                        }}
+                        options={{
+                            background: {
+                                color: {
+                                    value: "transparent"
+                                }
+                            },
+                            fpsLimit: 60,
+                            particles: {
+                                color: {
+                                    value: "#8590AA"
+                                },
+                                links: {
+                                    color: "#8590AA",
+                                    distance: 150,
+                                    enable: true,
+                                    opacity: 0.7,     // Aumentado de 0.5 a 0.7
+                                    width: 2.5        // Aumentado de 1 a 2.5
+                                },
+                                move: {
+                                    enable: true,
+                                    direction: "none",
+                                    outModes: {
+                                        default: "bounce"
+                                    },
+                                    random: false,
+                                    speed: 1,
+                                    straight: false
+                                },
+                                number: {
+                                    density: {
+                                        enable: true,
+                                        area: 800
+                                    },
+                                    value: 80
+                                },
+                                opacity: {
+                                    value: 0.8       // Aumentado de 0.5 a 0.8
+                                },
+                                shape: {
+                                    type: "circle"
+                                },
+                                size: {
+                                    value: { min: 3, max: 6 }  // Aumentado de {min: 1, max: 3} a {min: 3, max: 6}
+                                }
+                            },
+                            detectRetina: true
+                        }}
+                    />
+
         <Box className="boxContainer" sx={{ maxWidth: 400, mx: 'auto', p: 2 }}>
             <Link to="/home">
                 <img src="/logo512.png" alt="Logo" className="logoAplicacion" />
@@ -106,7 +176,10 @@ function Login() {
                     </Box>
                 </form>
             </div>
+
         </Box>
+
+        </>
     );
 }
